moved VR stuff to separate class
