+++ OWML.Patcher/OWML.Patcher.csproj
@@ -49,6 +49,7 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="VRFilePatcher.cs" />
     <Compile Include="BinaryPatcher.cs" />
     <Compile Include="OWPatcher.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
